+++ jest.config.js
@@ -1,7 +1,10 @@
-/** @type {import('jest').Config} */
+/** @type {import("jest").Config} */
 module.exports = {
-  preset: 'ts-jest',
-  testEnvironment: 'node',
-  testMatch: ['**/tests/**/*.test.ts'],
-  clearMocks: true
+  preset: "ts-jest",
+  testEnvironment: "node",
+  testMatch: ["**/tests/**/*.test.ts"],
+  clearMocks: true,
+  transform: {
+    "^.+\\.ts$": ["ts-jest", { diagnostics: { ignoreCodes: [151002] } }]
+  }
 };